Week 1, Lesson 3, Exercise 2: Node Modules: Callbacks and Error Handling
Objectives and Outcomes

In this exercise, you will learn about callbacks and error handling in Node applications. At the end of this exercise, you will be able to:

- Using Callbacks and error handling in Node applications
- Using external Node modules

Conclusions

In this exercise, you learnt about using Callbacks and error handling in Node applications. In addition you learnt about using external Node modules.

Changes to be committed:
	modified:   .gitignore
	new file:   node-examples/rectangle-2.js
	new file:   node-examples/solve-2.js
	new file:   node-examples/solve-3.js