Week 2, Lesson 1, Exercise 1: Express Generator
Objectives and Outcomes

In this exercise, you will use the Express generator to scaffold out an Express application. Thereafter you will modify the application to support REST API making use of the Node modules that you developed as part of the first assignment. At the end of this exercise, you will be able to:

- Generating an Express application using the express-generator
- Modify the Express application to support the REST API by adding routes

Conclusions

In this exercise, you learnt to scaffold out an Express application using the express-generator. Thereafter you were able to build a REST API server.

Changes to be committed:
	modified:   ../.gitignore
	new file:   app.js
	new file:   bin/www
	new file:   package.json
	new file:   public/aboutus.html
	new file:   public/index.html
	new file:   public/stylesheets/style.css
	new file:   routes/dishRouter.js
	new file:   routes/index.js
	new file:   routes/leaderRouter.js
	new file:   routes/promoRouter.js
	new file:   routes/users.js
	new file:   views/error.jade
	new file:   views/index.jade
	new file:   views/layout.jade