Week 2, Lesson 3, Exercise 1: Node and MongoDB Part 1
Objectives and Outcomes

In this exercise you will install the Node MongoDB driver module and configure your Node application to communicate with the MongoDB server. At the end of this exercise, you will be able to:

- Install and use the Node MongoDB driver
- Interact with the MongoDB database from a Node application

Conclusions

In this exercise you installed the Node MongoDB driver and used it to communicate with the MongoDB server from a Node application.

Changes to be committed:
	modified:   ../.gitignore
	new file:   package.json
	new file:   simpleserver.js