Week 2, Lesson 3, Exercise 1: Node and MongoDB Part 2
Objectives and Outcomes

In this exercise you will continue to explore communicating from your Node application to the MongoDB server. At the end of this exercise you will be able to:

- Develop a Node module containing some common MongoDB operations
- Use the Node module in your application and communicate with the MongoDB server

Conclusions

In this exercise you created a Node module to package some database operations, and then used the module to interact with the MongoDB server.

Changes to be committed:
	new file:   operations.js
	new file:   server.js