Week 2, Lesson 4, Exercise 1: Mongoose ODM
Objectives and Outcomes

In this exercise you will explore the Mongoose ODM and learn about creating schemas and interacting with the MongoDB database using Mongoose methods. At the end of this exercise, you will be able to:

- Install Mongoose ODM and connect to a MongoDB Server
- Create Mongoose Schemas
- Perform Database operations with Mongoose methods

Conclusions

In this exercise you learnt about the Mongoose module and used it to create schemas and interact with the MongoDB server.

Changes to be committed:
	modified:   ../.gitignore
	new file:   models/dishes-1.js
	new file:   models/dishes-3.js
	new file:   package.json
	new file:   server-1.js
	new file:   server-2.js
	new file:   server-3.js