Week 3, Lesson 1, Exercise 1: REST API with Express, MongoDB and Mongoose
Objectives and Outcomes

In this exercise, you will integrate the REST API server based on the Express framework that you implemented earlier, together with the Mongoose schema and models that you developed as part of Assignment 2 to create a full-fledged REST API server. At the end of this exercise, you will be able to:

- Develop a full-fledged REST API server with Express, MongoDB and Mongoose
- Serve up various REST API end points together with interaction with the MongoDB server.

Conclusions

In this exercise you developed a full-fledged REST API server with Express, Mongo and Mongoose.

Changes to be committed:
	modified:   ../.gitignore
	new file:   app.js
	new file:   bin/www
	new file:   models/dishes.js
	new file:   models/leadership.js
	new file:   models/promotions.js
	new file:   package.json
	new file:   public/stylesheets/style.css
	new file:   routes/dishRouter.js
	new file:   routes/index.js
	new file:   routes/leaderRouter.js
	new file:   routes/promoRouter.js
	new file:   routes/users.js
	new file:   views/error.jade
	new file:   views/index.jade
	new file:   views/layout.jade