Week 3, Lesson 3, Exercise 1: Using Cookies
Objectives and Outcomes

In this exercise you will examine the use of cookies for authentication.  The server will send a signed cookie to the client upon successful authentication, and expects the client to include the cookie with every subsequent request. At the end of this exercise, you will be able to:

- Set up your Express application to send signed cookies.
- Set up your Express application to parse the cookies in the header of the incoming request messages

Conclusions

In this exercise you examined the use of cookies for tracking authenticated users so that subsequent access to the server can be enabled without need for authentication.

Changes to be committed:
	modified:   package.json
	new file:   server-2.js