Week 3, Lesson 3, Exercise 2: Express Sessions
Objectives and Outcomes

In this exercise you will use Express sessions to track authenticated users so as to enable authenticated access to server resources. At the end of this exercise you will be able to:

- Set up your Express server to use Express sessions to track authenticated- users
- Enable clients to access secure resources on the server after authentication.

Conclusions

In this exercise you set up the Express server to use express-session to track authenticated users to provide access to secure resources.

Changes to be committed:
	modified:   ../.gitignore
	modified:   package.json
	new file:   server-3.js