Week 3, Lesson 4, Exercise 1: User Authentication with Passport
Objectives and Outcomes

In this exercise, you will explore user authentication with JSON web tokens and the Passport module. You will be able to control access to some routes within your REST server. At the end of this exercise, you will be able to:

- Use JSON web tokens for token-based user authentication
- Use Passport module together with passport-local and passport-local-mongoose for setting up local authentication within your server.

Conclusions

In this exercise you used token-based verification together with the Passport module to verify the authenticity of users and control access to routes in a REST API server.

Changes to be committed:
	modified:   ../.gitignore
	new file:   app.js
	new file:   bin/www
	new file:   config.js
	new file:   models/dishes.js
	new file:   models/leadership.js
	new file:   models/promotions.js
	new file:   models/user.js
	new file:   package.json
	new file:   public/stylesheets/style.css
	new file:   routes/dishRouter.js
	new file:   routes/index.js
	new file:   routes/leaderRouter.js
	new file:   routes/promoRouter.js
	new file:   routes/users.js
	new file:   routes/verify.js
	new file:   views/error.jade
	new file:   views/index.jade
	new file:   views/layout.jade