Week 4, Lesson 3, Exercise: Using OAuth with Passport and Facebook
Objectives and Outcomes

In this exercise you will make use of Passport OAuth support through the passport-facebook module together with Facebook's OAuth support to enable user authentication within your server. At the end of this exercise, you will be able to:

- Configure your server to support user authentication based on OAuth providers
- Use Passport OAuth support through the passport-facebook module to support OAuth based authentication with Facebook for your users.

Conclusions

In this exercise you learnt about using the Facebook OAuth support to enable authentication of your users and allowing them access to your server.

Changes to be committed:
	modified:   app.js
	new file:   authenticate.js
	modified:   config.js
	modified:   models/user.js
	modified:   package.json
	modified:   routes/users.js