Week 4, Lesson 4, Exercise 2: Loopback Data Sources and Access Control
Objectives and Outcomes

In this exercise, you will continue the exploration of Loopback. You will learn to set up a MongoDB as a data source and then set up access controls on the REST API endpoints. At the end of this exercise, you will be able to:

- Define data sources to be used by your Loopback server
- Set up access controls to various REST API end points.

Conclusions

In this exercise you learnt about setting up data sources and access control in a Loopback server.

Changes to be committed:
	new file:   common/models/customer.js
	new file:   common/models/customer.json
	modified:   common/models/dishes.json
	modified:   package.json
	new file:   server/boot/script.js
	modified:   server/datasources.json
	modified:   server/model-config.json