Week 4, Lesson 7, Exercise 3: Loopback Relations
Objectives and Outcomes

In this exercise you will explore the use of model relations in Loopback and how we can link various models by defining relations among them. In additionn you will explore the use of a timestamp mixin. At the end of this exercise, you will be able to:

- Define model relations among various Loopback models
- Make use of a mixin within the Loopback server

Conclusions

In this exercise you explored the use of model relations in Loopback and used a timestamp mixin.

Changes to be committed:
	new file:   common/models/comments.js
	new file:   common/models/comments.json
	modified:   common/models/customer.json
	modified:   common/models/dishes.json
	modified:   package.json
	modified:   server/model-config.json