Week 4, Lesson 6, Peer Graded Assignment: Assignment 4
Assignment Overview

In this assignment, you will be extending the router to support the ability to save and retrieve a list of favorite dishes by each of the registered users. All registered users in the system should have the ability to save any dish as their favorite dish, retrieve all their favorite dishes and remove one or all their favorite dishes. At the end of this assignment, your should have completed the following:

- Allowed users to select a dish as their favorite, and add it to the list of favorites that are saved on the server.
- Allowed users to retrieve the list of their favorite dishes from the server
- Delete one or all of their favorite dishes from their favorites list on the server.

Review criteria

Your assignment will be graded on the basis of the following review criteria:

- A new favoriteSchema and Favorites model has been correctly implemented to take advantage of Mongoose Population support to track the users and the list of favorite dishes using their ObjectIds in the favoriteSchema and Favorites model.
- The GET, POST and DELETE operations are well supported as per the specifications above
- The app.js has been updated to support the new route.

Changes to be committed:
	modified:   routes/favoriteRouter.js